Fix use of import/create app layout
