+++ run.py
@@ -1,5 +1,4 @@
 #!flask/bin/python
-from sbr_ui import create_application
+from sbr_ui import app
 
-app = create_application()
 app.run()
